Refactor CSV endpoints into new controller
