+++ conf/routes
@@ -19,6 +19,6 @@ GET /management/healthcheck controllers.Management.healthCheck
 DELETE /api/v1/services/ controllers.Management.deleteAllHistory
 DELETE /api/v1/services/:service controllers.Management.deleteSingleService(service: String)
 
-# generate csv
-GET /csv controllers.Management.generateCsvDownload()
-GET /csv-view controllers.Management.generateCsvString()
+# csv endpoints
+GET /csv controllers.CSVController.generateCsvDownload()
+GET /csv-view controllers.CSVController.generateCsvString()
